+++ Specs/DataSources/PolylineGeometryUpdaterSpec.js
@@ -91,6 +91,7 @@ defineSuite([
         expect(updater.isClosed).toBe(false);
         expect(updater.fillEnabled).toBe(false);
         expect(updater.fillMaterialProperty).toBe(undefined);
+        expect(updater.depthFailMaterialProperty).toBe(undefined);
         expect(updater.outlineEnabled).toBe(false);
         expect(updater.hasConstantFill).toBe(true);
         expect(updater.hasConstantOutline).toBe(true);
@@ -131,6 +132,7 @@ defineSuite([
         expect(updater.isClosed).toBe(false);
         expect(updater.fillEnabled).toBe(true);
         expect(updater.fillMaterialProperty).toEqual(new ColorMaterialProperty(Color.WHITE));
+        expect(updater.depthFailMaterialProperty).toBe(undefined);
         expect(updater.outlineEnabled).toBe(false);
         expect(updater.hasConstantFill).toBe(true);
         expect(updater.hasConstantOutline).toBe(true);
@@ -145,6 +147,13 @@ defineSuite([
         var updater = new PolylineGeometryUpdater(entity, scene);
         entity.polyline.material = new ColorMaterialProperty();
         expect(updater.fillMaterialProperty).toBe(entity.polyline.material);
+    });
+
+    it('Polyline depth fail material is correctly exposed.', function() {
+        var entity = createBasicPolyline();
+        var updater = new PolylineGeometryUpdater(entity, scene);
+        entity.polyline.depthFailMaterial = new ColorMaterialProperty();
+        expect(updater.depthFailMaterialProperty).toBe(entity.polyline.depthFailMaterial);
     });
 
     it('A time-varying positions causes geometry to be dynamic', function() {
@@ -198,6 +207,7 @@ defineSuite([
         var polyline = entity.polyline;
         polyline.show = new ConstantProperty(options.show);
         polyline.material = options.material;
+        polyline.depthFailMaterial = options.depthFailMaterial;
 
         polyline.width = new ConstantProperty(options.width);
         polyline.followSurface = new ConstantProperty(options.followSurface);
@@ -221,6 +231,11 @@ defineSuite([
         } else {
             expect(attributes.color).toBeUndefined();
         }
+        if (options.depthFailMaterial && options.depthFailMaterial instanceof ColorMaterialProperty) {
+            expect(attributes.depthFailColor.value).toEqual(ColorGeometryInstanceAttribute.toValue(options.depthFailMaterial.color.getValue(time)));
+        } else {
+            expect(attributes.depthFailColor).toBeUndefined();
+        }
         expect(attributes.show.value).toEqual(ShowGeometryInstanceAttribute.toValue(options.show));
         if (options.distanceDisplayCondition) {
             expect(attributes.distanceDisplayCondition.value).toEqual(DistanceDisplayConditionGeometryInstanceAttribute.toValue(options.distanceDisplayCondition));
@@ -237,10 +252,54 @@ defineSuite([
         });
     });
 
+    it('Creates expected per-color geometry with color depth fail appearance', function() {
+        validateGeometryInstance({
+            show : true,
+            material : new ColorMaterialProperty(Color.RED),
+            depthFailMaterial : new ColorMaterialProperty(Color.BLUE),
+            width : 3,
+            followSurface : false,
+            granularity : 1.0
+        });
+    });
+
+    it('Creates expected per-color geometry with material depth fail appearance', function() {
+        validateGeometryInstance({
+            show : true,
+            material : new ColorMaterialProperty(Color.RED),
+            depthFailMaterial : new GridMaterialProperty(),
+            width : 3,
+            followSurface : false,
+            granularity : 1.0
+        });
+    });
+
     it('Creates expected per-material geometry', function() {
         validateGeometryInstance({
             show : true,
             material : new GridMaterialProperty(),
+            width : 4,
+            followSurface : true,
+            granularity : 0.5
+        });
+    });
+
+    it('Creates expected per-material geometry with color depth fail appearance', function() {
+        validateGeometryInstance({
+            show : true,
+            material : new GridMaterialProperty(),
+            depthFailMaterial : new ColorMaterialProperty(Color.BLUE),
+            width : 4,
+            followSurface : true,
+            granularity : 0.5
+        });
+    });
+
+    it('Creates expected per-material geometry with color depth fail appearance', function() {
+        validateGeometryInstance({
+            show : true,
+            material : new GridMaterialProperty(),
+            depthFailMaterial : new GridMaterialProperty(),
             width : 4,
             followSurface : true,
             granularity : 0.5